Added very basic file interface for keyboard input and tty output

+++ include/cedos/drivers/keyboard.h
@@ -25,7 +25,7 @@ typedef struct {
      * Reads a single character from the keyboard
      * \return A single char corresponding to a key press.
      */
-    uint32_t (*read)(void);
+    uint8_t (*read)(void);
 } KB_DRIVER;
 
 //! PS/2 keyboard driver (default driver)
